# Conflicts:
#	KREST/KRESTManager.php
#	KREST/extensions/core.php
#	KREST/extensions/module.php
#	KREST/handlers/exceptionClasses.php
#	KREST/handlers/module.php
#	KREST/handlers/user.php
#	data/SugarBean.php
#	include/SpiceFTSManager/KREST/extensions/search.php
#	include/SpiceFTSManager/SpiceFTSHandler.php
#	include/SugarLogger/LogViewer.php
#	include/language/de_DE.lang.php
#	include/language/en_us.lang.php
#	include/utils.php
#	install/spicebeanguides/spicebeanguides.txt
#	metadata/system_krest.php
#	metadata/system_ui.php
#	modules/Administration/FTSDefault.php
#	modules/Calendar/KREST/handlers/krest.handler.calendar.php
#	modules/Calls/Call.php
#	modules/Emails/vardefs.php
#	modules/Mailboxes/Handlers/ImapHandler.php
#	modules/Mailboxes/vardefs.php
#	modules/Meetings/Meeting.php
#	modules/SystemUI/SpiceUIConfLoader.php
#	modules/SystemUI/SystemUIRESTHandler.php
#	modules/UserAbsences/UserAbsence.php
#	modules/UserAbsences/vardefs.php
#	modules/Users/User.php
#	modules/Users/vardefs.php
#	sugar_version.json
#	sugar_version.php
#	vendor/phenx/php-font-lib/README.md
#	vendor/phenx/php-font-lib/bower.json
#	vendor/phenx/php-font-lib/composer.json
#	vendor/phenx/php-font-lib/src/FontLib/AdobeFontMetrics.php
#	vendor/phenx/php-font-lib/src/FontLib/Autoloader.php
#	vendor/phenx/php-font-lib/src/FontLib/BinaryStream.php
#	vendor/phenx/php-font-lib/src/FontLib/EOT/File.php
#	vendor/phenx/php-font-lib/src/FontLib/EOT/Header.php
#	vendor/phenx/php-font-lib/src/FontLib/EncodingMap.php
#	vendor/phenx/php-font-lib/src/FontLib/Exception/FontNotFoundException.php
#	vendor/phenx/php-font-lib/src/FontLib/Font.php
#	vendor/phenx/php-font-lib/src/FontLib/Glyph/Outline.php
#	vendor/phenx/php-font-lib/src/FontLib/Glyph/OutlineComponent.php
#	vendor/phenx/php-font-lib/src/FontLib/Glyph/OutlineComposite.php
#	vendor/phenx/php-font-lib/src/FontLib/Glyph/OutlineSimple.php
#	vendor/phenx/php-font-lib/src/FontLib/Header.php
#	vendor/phenx/php-font-lib/src/FontLib/OpenType/File.php
#	vendor/phenx/php-font-lib/src/FontLib/OpenType/TableDirectoryEntry.php
#	vendor/phenx/php-font-lib/src/FontLib/Table/DirectoryEntry.php
#	vendor/phenx/php-font-lib/src/FontLib/Table/Table.php
#	vendor/phenx/php-font-lib/src/FontLib/Table/Type/cmap.php
#	vendor/phenx/php-font-lib/src/FontLib/Table/Type/glyf.php
#	vendor/phenx/php-font-lib/src/FontLib/Table/Type/head.php
#	vendor/phenx/php-font-lib/src/FontLib/Table/Type/hhea.php
#	vendor/phenx/php-font-lib/src/FontLib/Table/Type/hmtx.php
#	vendor/phenx/php-font-lib/src/FontLib/Table/Type/kern.php
#	vendor/phenx/php-font-lib/src/FontLib/Table/Type/loca.php
#	vendor/phenx/php-font-lib/src/FontLib/Table/Type/maxp.php
#	vendor/phenx/php-font-lib/src/FontLib/Table/Type/name.php
#	vendor/phenx/php-font-lib/src/FontLib/Table/Type/nameRecord.php
#	vendor/phenx/php-font-lib/src/FontLib/Table/Type/os2.php
#	vendor/phenx/php-font-lib/src/FontLib/Table/Type/post.php
#	vendor/phenx/php-font-lib/src/FontLib/TrueType/Collection.php
#	vendor/phenx/php-font-lib/src/FontLib/TrueType/File.php
#	vendor/phenx/php-font-lib/src/FontLib/TrueType/Header.php
#	vendor/phenx/php-font-lib/src/FontLib/TrueType/TableDirectoryEntry.php
#	vendor/phenx/php-font-lib/src/FontLib/WOFF/File.php
#	vendor/phenx/php-font-lib/src/FontLib/WOFF/Header.php
#	vendor/phenx/php-font-lib/src/FontLib/WOFF/TableDirectoryEntry.php
#	vendor/phenx/php-font-lib/tests/FontLib/FontTest.php
#	vendor/phenx/php-svg-lib/README.md
#	vendor/phenx/php-svg-lib/composer.json
#	vendor/phenx/php-svg-lib/phpunit.xml
#	vendor/phenx/php-svg-lib/src/Svg/DefaultStyle.php
#	vendor/phenx/php-svg-lib/src/Svg/Document.php
#	vendor/phenx/php-svg-lib/src/Svg/Gradient/Stop.php
#	vendor/phenx/php-svg-lib/src/Svg/Style.php
#	vendor/phenx/php-svg-lib/src/Svg/Surface/CPdf.php
#	vendor/phenx/php-svg-lib/src/Svg/Surface/SurfaceCpdf.php
#	vendor/phenx/php-svg-lib/src/Svg/Surface/SurfaceGmagick.php
#	vendor/phenx/php-svg-lib/src/Svg/Surface/SurfaceInterface.php
#	vendor/phenx/php-svg-lib/src/Svg/Surface/SurfacePDFLib.php
#	vendor/phenx/php-svg-lib/src/Svg/Tag/AbstractTag.php
#	vendor/phenx/php-svg-lib/src/Svg/Tag/Anchor.php
#	vendor/phenx/php-svg-lib/src/Svg/Tag/Circle.php
#	vendor/phenx/php-svg-lib/src/Svg/Tag/ClipPath.php
#	vendor/phenx/php-svg-lib/src/Svg/Tag/Ellipse.php
#	vendor/phenx/php-svg-lib/src/Svg/Tag/Group.php
#	vendor/phenx/php-svg-lib/src/Svg/Tag/Image.php
#	vendor/phenx/php-svg-lib/src/Svg/Tag/Line.php
#	vendor/phenx/php-svg-lib/src/Svg/Tag/LinearGradient.php
#	vendor/phenx/php-svg-lib/src/Svg/Tag/Path.php
#	vendor/phenx/php-svg-lib/src/Svg/Tag/Polygon.php
#	vendor/phenx/php-svg-lib/src/Svg/Tag/Polyline.php
#	vendor/phenx/php-svg-lib/src/Svg/Tag/RadialGradient.php
#	vendor/phenx/php-svg-lib/src/Svg/Tag/Rect.php
#	vendor/phenx/php-svg-lib/src/Svg/Tag/Shape.php
#	vendor/phenx/php-svg-lib/src/Svg/Tag/Stop.php
#	vendor/phenx/php-svg-lib/src/Svg/Tag/StyleTag.php
#	vendor/phenx/php-svg-lib/src/Svg/Tag/Text.php
#	vendor/phenx/php-svg-lib/src/Svg/Tag/UseTag.php
#	vendor/phenx/php-svg-lib/src/autoload.php
#	vendor/phenx/php-svg-lib/tests/Svg/StyleTest.php

+++ modules/SystemUI/SpiceUIConfLoader.php
@@ -192,14 +192,17 @@ class SpiceUIConfLoader
         foreach ($response as $tb => $content) {
             //truncate command
             $tables[] = $tb;
+            $thisCols = $this->getTableColumns($tb);
             switch ($tb) {
                 case 'syslangs':
                     $delQ = "DELETE FROM $tb WHERE 1=1";//$GLOBALS['db']->truncateTableSQL($tb);
                     break;
                 default:
-                    $delQ = "DELETE FROM $tb WHERE package IN('" . implode("','", $params['packages']) . "') ";
-                    if (in_array('core', $params['packages']))
-                        $delQ .= "OR package IS NULL OR package=''";
+                    if(array_search('package', $thisCols) !== false) {
+                        $delQ = "DELETE FROM $tb WHERE package IN('" . implode("','", $params['packages']) . "') ";
+                        if (in_array('core', $params['packages']))
+                            $delQ .= "OR package IS NULL OR package=''";
+                    }
             }
 
             $truncates[] = $delQ;
@@ -212,16 +215,15 @@ class SpiceUIConfLoader
                         "<br/>Action aborted");
 
                 //prcess only selected packages and empty package values
-                if (empty($decodeData['package']) && !in_array('core', $params['packages'])) {
+                if (isset($decodeData['package']) && empty($decodeData['package']) && !in_array('core', $params['packages'])) {
                     continue;
-                } elseif (!empty($decodeData['package']) && !in_array($decodeData['package'], $params['packages'])) {
+                } elseif (isset($decodeData['package']) && !empty($decodeData['package']) && !in_array($decodeData['package'], $params['packages'])) {
                     continue;
                 }
 
                 //compare table column names
                 if (!$tbColCheck) {
                     $referenceCols = array_keys($decodeData);
-                    $thisCols = $this->getTableColumns($tb);
                     if (!empty(array_diff($referenceCols, $thisCols))) {
                         die("Table structure for $tb is not up-to-date." .
                             "<br/>Reference table = " . implode(", ", $referenceCols) .
